requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
blake3==0.4.1
cryptography==43.0.3
pycryptodome==3.20.0
base58==2.1.1
